+++ backend/spring/src/main/java/at/gradwohl/website/controller/FirmaController.java
@@ -1,7 +1,11 @@
 package at.gradwohl.website.controller;
 
+import at.gradwohl.website.config.JwtService;
 import at.gradwohl.website.model.firma.Firma;
 import at.gradwohl.website.service.firma.FirmaService;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,23 +14,31 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
+@RequiredArgsConstructor
+@SecurityRequirement(name ="jwt-auth")
 @RequestMapping(path = "firma")
 public class FirmaController {
 
     private final FirmaService firmaService;
-
-    @Autowired
-    public FirmaController(FirmaService firmaService) {
-        this.firmaService = firmaService;
-    }
+    private final JwtService jwtService;
 
     @GetMapping
-    public List<Firma> getAllFirma() {
-        return firmaService.getAllFirma();
+    public ResponseEntity<List<Firma>> getAllFirma(HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        return new ResponseEntity<>(firmaService.getAllFirma(), HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<Firma> addFirma(@RequestBody Firma firma) {
+    public ResponseEntity<Firma> addFirma(
+            @RequestBody Firma firma,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         Firma newFirma = firmaService.addFirma(firma);
         return new ResponseEntity<>(newFirma, HttpStatus.CREATED);
     }
@@ -34,7 +46,12 @@ public class FirmaController {
     @PutMapping("/{id}")
     public ResponseEntity<Firma> updateFirma(
             @PathVariable("id") String id,
-            @RequestBody Firma updatedFirma) {
+            @RequestBody Firma updatedFirma,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         Firma result = firmaService.updateFirma(id, updatedFirma);
         if (result != null) {
             return ResponseEntity.ok(result);
@@ -43,7 +60,13 @@ public class FirmaController {
         }
     }
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteFirma(@PathVariable("id") String id) {
+    public ResponseEntity<Void> deleteFirma(
+            @PathVariable("id") String id,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         firmaService.deleteFirma(id);
         return ResponseEntity.noContent().build();
     }
