+++ backend/spring/src/main/java/at/gradwohl/website/controller/ProduktController.java
@@ -1,7 +1,11 @@
 package at.gradwohl.website.controller;
 
+import at.gradwohl.website.config.JwtService;
 import at.gradwohl.website.model.produkt.Produkt;
 import at.gradwohl.website.service.produkt.ProduktService;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,16 +14,22 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
+@RequiredArgsConstructor
+@SecurityRequirement(name ="jwt-auth")
 @RequestMapping(path = "produkt")
 public class ProduktController {
 
     private final ProduktService produktService;
-
-    @Autowired
-    public ProduktController(ProduktService produktService) { this.produktService = produktService; }
+    private final JwtService jwtService;
 
     @PostMapping
-    public ResponseEntity<Produkt> insertProduct(@RequestBody Produkt produkt) {
+    public ResponseEntity<Produkt> insertProduct(
+            @RequestBody Produkt produkt,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         try {
             Produkt insertedProdukt = produktService.insertProduct(produkt);
             return ResponseEntity.ok(insertedProdukt);
@@ -29,7 +39,14 @@ public class ProduktController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Produkt> updateProduct(@PathVariable int id, @RequestBody Produkt updatedProdukt) {
+    public ResponseEntity<Produkt> updateProduct(
+            @PathVariable int id,
+            @RequestBody Produkt updatedProdukt,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         try {
             Produkt updatedProduct = produktService.updateProduct(id, updatedProdukt);
             return ResponseEntity.ok(updatedProduct);
@@ -39,7 +56,13 @@ public class ProduktController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteProduktById(@PathVariable int id) {
+    public ResponseEntity<String> deleteProduktById(
+            @PathVariable int id,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         try {
             produktService.deleteProduktById(id);
             return ResponseEntity.ok("Produkt deleted");
@@ -51,7 +74,13 @@ public class ProduktController {
     }
 
     @GetMapping("/ware/{id}")
-    public ResponseEntity<Produkt> getProductById(@PathVariable int id) {
+    public ResponseEntity<Produkt> getProductById(
+            @PathVariable int id,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsVerkauf(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         try{
             return ResponseEntity.ok(produktService.getProduktById(id));
         } catch (IllegalArgumentException e) {
@@ -60,26 +89,46 @@ public class ProduktController {
     }
 
     @GetMapping("/{produktgruppeName}")
-    public List<Produkt> getProductsByProduktgruppe(@PathVariable String produktgruppeName) {
-        return produktService.getProductsByProduktgruppeName(produktgruppeName);
+    public ResponseEntity<List<Produkt>> getProductsByProduktgruppe(
+            @PathVariable String produktgruppeName,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsVerkauf(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        return new ResponseEntity<>(produktService.getProductsByProduktgruppeName(produktgruppeName), HttpStatus.OK);
     }
 
     @GetMapping("/{produktgruppeName}/{hb}")
-    public List<Produkt> getProductsByProduktgruppe(
+    public ResponseEntity<List<Produkt>> getProductsByProduktgruppe(
             @PathVariable String produktgruppeName,
-            @PathVariable boolean hb) {
-        return produktService.getProduktbyGruppeAndHB(produktgruppeName, hb);
+            @PathVariable boolean hb,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsVerkauf(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        return new ResponseEntity<>(produktService.getProduktbyGruppeAndHB(produktgruppeName, hb), HttpStatus.OK);
     }
 
     @GetMapping("/hb/{hb}")
-    public List<Produkt> getProductsByProduktgruppe(
-            @PathVariable boolean hb) {
-        return produktService.getHBProdukte(hb);
+    public ResponseEntity<List<Produkt>> getProductsByProduktgruppe(
+            @PathVariable boolean hb,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsVerkauf(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        return new ResponseEntity<>(produktService.getHBProdukte(hb), HttpStatus.OK);
     }
 
     @GetMapping
-    public  List<Produkt> getAllProducts(){
-        return produktService.getAllProdukts();
+    public  ResponseEntity<List<Produkt>> getAllProducts(HttpServletRequest request){
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsVerkauf(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        return new ResponseEntity<>(produktService.getAllProdukts(), HttpStatus.OK);
     }
 
 }
