+++ backend/spring/src/main/java/at/gradwohl/website/controller/RoleController.java
@@ -1,7 +1,11 @@
 package at.gradwohl.website.controller;
 
+import at.gradwohl.website.config.JwtService;
 import at.gradwohl.website.model.mitarbeiterrole.MitarbeiterRole;
 import at.gradwohl.website.service.role.RoleService;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,23 +14,31 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
+@RequiredArgsConstructor
+@SecurityRequirement(name ="jwt-auth")
 @RequestMapping(path = "role")
 public class RoleController {
 
     private final RoleService roleService;
-
-    @Autowired
-    public RoleController(RoleService roleService){
-        this.roleService = roleService;
-    }
+    private final JwtService jwtService;
 
     @GetMapping
-    public List<MitarbeiterRole> getAllRole() {
-        return roleService.getAllRole();
+    public ResponseEntity<List<MitarbeiterRole>> getAllRole(HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsVerkauf(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
+        return new ResponseEntity<>(roleService.getAllRole(), HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<MitarbeiterRole> addRole(@RequestBody MitarbeiterRole role) {
+    public ResponseEntity<MitarbeiterRole> addRole(
+            @RequestBody MitarbeiterRole role,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         MitarbeiterRole newRole = roleService.addRole(role);
         return new ResponseEntity<>(newRole, HttpStatus.CREATED);
     }
@@ -34,7 +46,12 @@ public class RoleController {
     @PutMapping("/{id}")
     public ResponseEntity<MitarbeiterRole> updateRole(
             @PathVariable("id") String id,
-            @RequestBody MitarbeiterRole updatedRole) {
+            @RequestBody MitarbeiterRole updatedRole,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         MitarbeiterRole result = roleService.updateRole(id, updatedRole);
         if (result != null) {
             return ResponseEntity.ok(result);
@@ -43,7 +60,13 @@ public class RoleController {
         }
     }
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteRole(@PathVariable("id") String id) {
+    public ResponseEntity<Void> deleteRole(
+            @PathVariable("id") String id,
+            HttpServletRequest request) {
+        String myHeader = request.getHeader("Authorization").substring(7);
+        if(!jwtService.getRoleIsZentrale(myHeader))
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         roleService.deleteRole(id);
         return ResponseEntity.noContent().build();
     }
